Refactor the codebase (#213)
* refactor the codebase.

* rename recent contributors utility file.
